1. Changes in order details UI according Paid & UnPaid order 2. Unpaid & Paid order show in order list 3. some major changes done

+++ app/src/main/java/com/vasyerp/selfcheckout/models/orderlist/OrderModel.java
@@ -16,7 +16,7 @@ import lombok.NoArgsConstructor;
 public class OrderModel {
     @SerializedName("salesId")
     @Expose
-    private int salesId;
+    private long salesId;
     @SerializedName("total")
     @Expose
     private float total;
